Enable possibility of Aarch64 build
Tizen toolchain team in Samsung company is planning to build Tizen
for new arm 64 bit architecture - Aarch64. But package
"leveldb" cannot be built for this architecture because of build errors
related with the fact that for architecture "aarch64" gcc compiler defines
the macro that is different from the macro defined for standard 32-bit ARM
architecture. That's why we have added the check of __aarch64__ macro in
one header file.

Change-Id: I7be468b49c8e1a552c1c7e52d6c08166e1463570
Signed-off-by: Ilya Palachev <i.palachev@samsung.com>