Enable possibility of Aarch64 build
Package "leveldb" cannot be built for architecture "aarch64" because
of build errors related with the fact that for architecture
"aarch64" gcc compiler defines the macro that is different from the
macro defined for standard 32-bit ARM architecture. That's why we
have added the check of __aarch64__ macro in one header file.

Change-Id: I7be468b49c8e1a552c1c7e52d6c08166e1463570
Signed-off-by: Ilya Palachev <i.palachev@samsung.com>